chapter-40-deployment-patterns/code/requirements.txt
# requirements.txt
#
# Chapter 40: Deployment Patterns
#
# Install with: pip install -r requirements.txt
# Or with uv: uv pip install -r requirements.txt

# Core dependencies
anthropic>=0.18.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Optional: Redis for caching and task queues
# redis>=5.0.0
# rq>=1.15.0

# Optional: System monitoring
# psutil>=5.9.0

# Development/Testing (uncomment as needed)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0
